Factors out Node protocol to typing module and adds import shortcuts.
