+++ example.py
@@ -1,9 +1,8 @@
 import random
 
-from sherwood.events.animator import tree_renderer
+from sherwood.events import tree_renderer
 from sherwood.renderers import draw_avl_tree, draw_redblack_tree
-from sherwood.trees.avl import AVLTree
-from sherwood.trees.redblack import RedBlackTree
+from sherwood.trees import AVLTree, RedBlackTree
 
 
 def main():
