+++ src/sherwood/events/__init__.py
@@ -1,4 +1,4 @@
 from .animator import Animator, tree_renderer
-from .base import Bus, Event
+from .base import AnimationNode, Bus, Event
 
-__all__ = "Animator", "Bus", "Event", "tree_renderer"
+__all__ = "AnimationNode", "Animator", "Bus", "Event", "tree_renderer"
