+++ src/sherwood/renderers/__init__.py
@@ -20,3 +20,6 @@ draw_redblack_tree: Renderer = TreeRenderer(
     edge_attr_funcs=[color_marked_node_edge],
     node_attr_funcs=[outline_marked_node, red_black_node_color],
 )
+
+
+__all__ = "Renderer", "draw_tree", "draw_avl_tree", "draw_redblack_tree"
