+++ src/sherwood/trees/__init__.py
@@ -1,0 +1,5 @@
+from .avl import AVLNode, AVLTree
+from .base import BinaryNode, Tree
+from .redblack import RBNode, RedBlackTree
+
+__all__ = "AVLNode", "AVLTree", "BinaryNode", "RBNode", "RedBlackTree", "Tree"
